+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Description TBD
 * Clear search text whenever filters are applied.
 * Sort weapon filters by name.
 * Sort weapon archetype filters by RPM/charge time if applicable.
+* Show Crafted bonus option now hidden for non-crafted weapons.
 
 ## **Improve Page Load Times + Quality of Life** [ *1.1.0* ] 2023-02-13
 This change brings improvements to page load times as well as some quality of life changes such as